server/requirements.txt
fastapi
uvicorn
python-multipart
mysql-connector-python
pydantic
pyx12
google-cloud-storage
xmltodict
airefinery-sdk>=1.29.0
python-dotenv>=1.0.0
pymongo>=4.6.0
